fix: history: Fixed history (again)
This time around I noticed that when I used the history the workspace
disappeared. This was due to my previous choice of setting the
`excludeFromExport` flag on it - because I don't want it in my SVGs.

Turns out that in this case the history gets borked, because it uses an
export feature…

I've now build my own excludeFromExport feature by changing the
export/save code. So yeah

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>