fix: workspace: Use 'before:render' event
Previously the 'after:render' event was used to trigger workspace
changes like dynamic resizing. As it turns out this is the better event
to use to make these changes, as one can directly influence the
rendering process which removes stuttering seeming changes.

For this change to work `renderAll` in `setSize` is only allowed to
trigger when the canvas is allowed to. Luckily this can be easily
controlled by changing the `canvas.renderOnAddRemove` property.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>