feat: editing: Add automatic ordering / align
With this patch objects can be automatically aligned / ordered on the
workspace. This is an important features for label makers, as sometimes
one just wants to plug multiple elements on the UI which are centered
and positioned behind each other.

The order follows the UI element order (top most element is last), which
is a property that can already be influenced by fabric.js.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>